[REF] split file by objects
