+++ l10n_br_delivery/models/__init__.py
@@ -1,4 +1,7 @@
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
 from . import carrier
+from . import vehicle
+from . import shipment
 from . import sale_order
 from . import stock_picking
-from . import l10n_br_delivery
